Flubbed refactor
